refactor: #52 minor fix
