+++ src/main/resources/templates/main-home.html
@@ -205,7 +205,6 @@
     window.pageType = "home";
 </script>
 <script type="module" src="../js/sse.js"></script>
-<script src="../js/my.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+<script src="../js/my.js" defer></script>
 </body>
 </html>
